requirements.txt
flask
python-dotenv
requests
authlib
six
flask_sqlalchemy
flask-cors
pyjwt